+++ lib/src/assets_audio_player.dart
@@ -11,6 +11,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:rxdart/rxdart.dart';
+import 'package:rxdart/subjects.dart';
 import 'package:uuid/uuid.dart';
 
 import 'applifecycle.dart';
@@ -89,10 +90,8 @@ class PlayerEditor {
   void onAudioReplacedAt(int index, bool keepPlayingPositionIfCurrent) {
     assetsAudioPlayer._updatePlaylistIndexes();
     if (assetsAudioPlayer._playlist!.playlistIndex == index) {
-      final currentPosition =
-          assetsAudioPlayer.currentPosition.valueWrapper?.value;
-      final isPlaying =
-          assetsAudioPlayer.isPlaying.valueWrapper?.value ?? false;
+      final currentPosition = assetsAudioPlayer.currentPosition.valueOrNull;
+      final isPlaying = assetsAudioPlayer.isPlaying.valueOrNull ?? false;
       //print('onAudioReplacedAt/ currentPosition : $currentPosition');
       if (keepPlayingPositionIfCurrent && currentPosition != null) {
         assetsAudioPlayer._openPlaylistCurrent(
@@ -429,7 +428,7 @@ class AssetsAudioPlayer {
   /// returns the looping state : true -> looping, false -> not looping
   LoopMode? get currentLoopMode => _loopMode.value;
 
-  bool get shuffle => _shuffle.valueWrapper?.value ?? false;
+  bool get shuffle => _shuffle.valueOrNull ?? false;
 
   bool _stopped = false;
 
@@ -570,7 +569,7 @@ class AssetsAudioPlayer {
           break;
         case METHOD_CURRENT:
           if (call.arguments == null) {
-            final current = _current.valueWrapper?.value;
+            final current = _current.valueOrNull;
             if (current != null) {
               final finishedPlay = Playing(
                 audio: current.audio,
@@ -666,8 +665,7 @@ class AssetsAudioPlayer {
             pause();
             break;
           case PlayInBackground.disabledRestoreOnForeground:
-            _wasPlayingBeforeEnterBackground =
-                isPlaying.valueWrapper?.value ?? false;
+            _wasPlayingBeforeEnterBackground = isPlaying.valueOrNull ?? false;
             pause();
             break;
         }
@@ -741,10 +739,9 @@ class AssetsAudioPlayer {
   Future<bool> previous({bool keepLoopMode = true}) async {
     if (_playlist != null) {
       // more than 5 sec played, go back to the start of audio
-      if (_currentPosition.valueWrapper?.value != null) {
-        if (_currentPosition.valueWrapper!.value.inSeconds >= 5) {
-          await seek(Duration.zero, force: true);
-        }
+      if (_currentPosition.valueOrNull != null &&
+          _currentPosition.valueOrNull!.inSeconds >= 5) {
+        await seek(Duration.zero, force: true);
       } else if (_playlist!.hasPrev()) {
         if (!keepLoopMode) {
           if (loopMode.value == LoopMode.single) {
@@ -764,7 +761,7 @@ class AssetsAudioPlayer {
   }
 
   void _onPositionReceived(dynamic argument) {
-    final oldValue = _currentPosition.valueWrapper?.value;
+    final oldValue = _currentPosition.valueOrNull;
     int? newValue;
     if (argument is int) {
       final value = argument;
@@ -861,7 +858,7 @@ class AssetsAudioPlayer {
         }
       }
       if (_playlist!.hasNext()) {
-        final curr = _current.valueWrapper?.value;
+        final curr = _current.valueOrNull;
         if (curr != null) {
           _playlistAudioFinished.add(Playing(
             audio: curr.audio,
@@ -876,7 +873,7 @@ class AssetsAudioPlayer {
         return true;
       } else if (loopMode.value == LoopMode.playlist) {
         //last element
-        final curr = _current.valueWrapper?.value;
+        final curr = _current.valueOrNull;
         if (curr != null) {
           _playlistAudioFinished.add(Playing(
             audio: curr.audio,
@@ -895,7 +892,7 @@ class AssetsAudioPlayer {
         return true;
       } else if (requestByUser) {
         //last element
-        final curr = _current.valueWrapper?.value;
+        final curr = _current.valueOrNull;
         if (curr != null) {
           _playlistAudioFinished.add(Playing(
             audio: curr.audio,
@@ -1043,15 +1040,13 @@ class AssetsAudioPlayer {
           'headPhoneStrategy': describeHeadPhoneStrategy(_headPhoneStrategy),
           'audioFocusStrategy': describeAudioFocusStrategy(_audioFocusStrategy),
           'displayNotification': _showNotification,
-          'volume': forcedVolume ?? volume.valueWrapper?.value ?? defaultVolume,
+          'volume': forcedVolume ?? volume.valueOrNull ?? defaultVolume,
           'playSpeed': playSpeed ??
               audio.playSpeed ??
-              this.playSpeed.valueWrapper?.value ??
+              this.playSpeed.valueOrNull ??
               defaultPlaySpeed,
-          'pitch': pitch ??
-              audio.pitch ??
-              this.pitch.valueWrapper?.value ??
-              defaultPitch,
+          'pitch':
+              pitch ?? audio.pitch ?? this.pitch.valueOrNull ?? defaultPitch,
         };
         if (seek != null) {
           params['seek'] = seek.inMilliseconds.round();
@@ -1256,7 +1251,7 @@ class AssetsAudioPlayer {
   ///     _assetsAudioPlayer.playOfPause();
   ///
   Future<void> playOrPause() async {
-    final playing = _isPlaying.valueWrapper?.value ?? true;
+    final playing = _isPlaying.valueOrNull ?? true;
     if (playing) {
       await pause();
     } else {
@@ -1360,12 +1355,11 @@ class AssetsAudioPlayer {
   ///
   Future<void> seekBy(Duration by) async {
     // only if playing a song
-    final playing = current.valueWrapper?.value;
+    final playing = current.valueOrNull;
     if (playing != null) {
       final totalDuration = playing.audio.duration;
 
-      final currentPosition =
-          this.currentPosition.valueWrapper?.value ?? Duration();
+      final currentPosition = this.currentPosition.valueOrNull ?? Duration();
 
       if (by.inMilliseconds >= 0) {
         final nextPosition = currentPosition + by;
@@ -1379,8 +1373,7 @@ class AssetsAudioPlayer {
         await seek(currentPositionCapped);
       } else {
         // only if playing a song
-        final currentPosition =
-            this.currentPosition.valueWrapper?.value ?? Duration();
+        final currentPosition = this.currentPosition.valueOrNull ?? Duration();
         final nextPosition = currentPosition + by;
 
         // don't seek less that 0
@@ -1410,6 +1403,7 @@ class AssetsAudioPlayer {
   /// Tells the media player to stop the current song, then release the MediaPlayer
   ///     _assetsAudioPlayer.stop();
   ///
+
   Future<void> stop({bool sendCallback = true}) async {
     _sendCallbackonFinished = sendCallback;
     return _stop(removeNotification: true);